+++ photocolle-sdk/GHUnitTests/DCAuthenticationContextKeyChainTest.m
@@ -2,7 +2,7 @@
 
 #import "DCAuthenticationContext_Private.h"
 #import "DCOAuth2Authentication.h"
-#import "GTMHTTPFetcherService.h"
+#import "GTMSessionFetcherService.h"
 #import "DCErrors.h"
 #import "DCTestUtils.h"
 #import "DCTypeRefHolder.h"
