+++ photocolle-sdk/photocolle-sdk/DCAuthority.m
@@ -7,7 +7,7 @@
 #import "DCOAuth2ViewController.h"
 #import "DCErrorUtils.h"
 #import "DCOAuth2Authentication.h"
-#import "DCHTTPFetcherService.h"
+#import "DCSessionFetcherService.h"
 
 #import "GTMOAuth2ViewControllerTouch.h"
 #import "GTMOAuth2SignIn.h"
@@ -339,7 +339,7 @@ static NSString * const DEFAULT_TOKEN_URL =
     BOOL isTimeout = dispatch_semaphore_wait(semaphore,
             dispatch_time(DISPATCH_TIME_NOW, 60 * NSEC_PER_SEC )) == 0
             ? NO : YES;
-    [context.fetcherService clearHistory];
+    [(DCSessionFetcherService *)context.fetcherService clearHistory];
     if (outError != nil) {
         *error = outError;
         NSLog(@"refresh token was failed: %@", [outError description]);
