+++ photocolle-sdk/photocolle-sdk/DCSessionFetcherService_Private.h
@@ -1,4 +1,4 @@
-#import "DCHTTPFetcherService.h"
+#import "DCSessionFetcherService.h"
 
 @interface DCSessionFetcherService ()
