+++ photocolle-sdk/photocolle-sdk/ext/gtm/OAuth2/GTMOAuth2Authentication.m
@@ -15,7 +15,7 @@
 
 /* 2015 Kii corp.
  *
- * Prefixes are changed from GTM to DCDCGTM.
+ * Prefixes are changed from GTM to DCGTM.
  *
  * Targets of changing prefix are all classes, protocols, extensions,
  * categoriesconst values, comments and etc.
@@ -731,7 +731,7 @@ finishedRefreshWithFetcher:(DCGTMOAuth2Fetcher *)fetcher
     appID = [appID stringByAppendingFormat:@"/%@", version];
   }
 
-  NSString *userAgent = @"DCGTM-oauth2";
+  NSString *userAgent = @"gtm-oauth2";
   if (appID) {
     userAgent = [userAgent stringByAppendingFormat:@" %@", appID];
   }
