+++ photocolle-sdk/photocolle-sdk/ext/gtm/OAuth2/GTMOAuth2ViewControllerTouch.m
@@ -15,7 +15,7 @@
 
 /* 2015 Kii corp.
  *
- * Prefixes are changed from GTM to DCDCGTM.
+ * Prefixes are changed from GTM to DCGTM.
  *
  * Targets of changing prefix are all classes, protocols, extensions,
  * categoriesconst values, comments and etc.
